moved entry target test from transition to region validation
